requirements.txt
asgiref==3.8.1
attrs==25.1.0
boto3==1.35.24
botocore==1.35.99
coverage==7.6.12
Django==5.0
django-admin-list-filter-dropdown==1.0.3
django-autocomplete-light==3.11.0
django-cleanup==9.0.0
django-cors-headers==4.4.0
django-filter==24.3
django-phonenumber-field==8.0.0
django-storages==1.14.4
django-timezone-field==7.0
django-tz-detect==0.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
phonenumbers==8.13.54
pillow==10.4.0
pluggy==1.5.0
PyJWT==2.10.1
pytest==8.3.3
pytest-cov==5.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
s3transfer==0.10.4
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wheel==0.45.1
